refactor: rename metric to telemetry
